Use explicit formats when calling api
From http://weblog.rubyonrails.org/2011/4/21/jquery-new-default

"In particular, if your application uses jQuery the default Accept
header in the test will be different from the Accept header sent by
jQuery. That's some vinegar: Rails encourages you to base interfaces on
explicit format parameters. So, for example, best practice is to provide
a /users.json endpoint, rather than a /users endpoint with a JSON
representation chosen depending on the Accept header."

So lets not go off the rails golden path when it's easy to stay on it