Handle I18n::Backend::Fallbacks
The problem was that by the time `default` was called "for real" it gets called without the `object` parameter. It gets called once with `options[:fallback] = true` and a valid `object`, so, without the patch, the code fails:

```
  1) I18n::Backend::Base#default records missing translations when fallbacks are enabled
     Failure/Error: i18n.translate('en', 'object', :default => 'subject')
       #<Localeapp::MissingTranslations:0x007f8403ae9190> received :add with unexpected arguments
         expected: ("en", "object", "subject", {:default=>"subject"})
              got: ("en", nil, nil, {:default=>"subject"})
     # ./lib/localeapp/default_value_handler.rb:9:in `default'
     # ./spec/localeapp/default_value_handler_spec.rb:74:in `block (3 levels) in <top (required)>'
     # ./spec/localeapp/default_value_handler_spec.rb:72:in `block (2 levels) in <top (required)>'
```

Obviously `nil` is not the value that was intended.

These are the calls to default:

```ruby
i18n.translate(:en, 'my.object', :default => 'my default')

call to default(locale, object, subject, options)
:en
"my.object"
[]
{:default=>[], :fallback=>true}

call to default(locale, object, subject, options)
"en"
nil
"my default"
{:default=>"my default"}
```

As you can see, by the time the subject is the default, the object is nil.