Stop checking project and app default locale at install
  Any project created with a different default locale than `en' will
make the installer output a warning: we cannot know the default locale
configured for the current rails app, we would need to boot it.

  Doing the check at runtime is complex, we would need to rely on a hook
like `Rails.application.config.after_initialize', and at this time we
don't know the default locale for the Localeapp project. Requesting the
API at this time could prevent the app from booting correctly.