Fix a TypeMismatch error in ExceptionHandler
This bug happens when blacklisted keys pattern is nil
(_i.e._ when the user has not configured it in its application).

It's due to `String#match?` not handling `nil` as parameter.

We fix it by using `String#=~` instead, which does handle nil as
parameter (and considers it never matching, which is what we want).

This bug was introduced in 4e46de6080f3810be7ea76868f228b4d2d11afaa